// fe/src/types/UserData.ts
export type UserData = {
  id: number;
  first_name: string;
  last_name: string;
  email: string;
  status: number;
  dateOfBirth?: string;
  totalFollower?: number;
  totalFollowing?: number;
  facebookLink?: string;
  discordLink?: string;
  githubLink?: string;
  maxim?: string;
  career?: string;
  flagFirst: boolean;
  avatarPath?: string;
  createdAt: string;
  updatedAt: string;
  role: string;
  gender: number;
}